+++ models/metric_java.py
@@ -1,0 +1,56 @@
+from models.metric_common import MetricCommon
+
+class MetricJava(MetricCommon):
+
+    def __init__(self):
+        # call init metriccommon
+        # query metric ck
+
+    # CK metrics
+    ck_cbo = Column(Float)
+    ck_cbo_modified = Column(Float)
+    ck_fan_in = Column(Float)
+    ck_fan_out = Column(Float)
+    ck_dit = Column(Float)
+    ck_noc = Column(Float)
+    ck_nom = Column(Float)
+    ck_nopm = Column(Float)
+    ck_noprm = Column(Float)
+    ck_num_fields = Column(Float)
+    ck_num_methods = Column(Float)
+    ck_num_visible_methods = Column(Float)
+    ck_nosi = Column(Float)
+    ck_rfc = Column(Float)
+    ck_wmc = Column(Float)
+    ck_loc = Column(Float)
+    ck_lcom = Column(Float)
+    ck_lcom_modified = Column(Float)
+    ck_tcc = Column(Float)
+    ck_lcc = Column(Float)
+    ck_qty_returns = Column(Float)
+    ck_qty_loops = Column(Float)
+    ck_qty_comparisons = Column(Float)
+    ck_qty_try_catch = Column(Float)
+    ck_qty_parenth_exps = Column(Float)
+    ck_qty_str_literals = Column(Float)
+    ck_qty_numbers = Column(Float)
+    ck_qty_math_operations = Column(Float)
+    ck_qty_math_variables = Column(Float)
+    ck_qty_nested_blocks = Column(Float)
+    ck_qty_ano_inner_cls_and_lambda = Column(Float)
+    ck_qty_unique_words = Column(Float)
+    ck_numb_log_stmts = Column(Float)
+    ck_has_javadoc = Column(Float)
+    ck_modifiers = Column(Float)
+    ck_usage_vars = Column(Float)
+    ck_usage_fields = Column(Float)
+    ck_method_invok = Column(Float)
+
+    # JPeek metrics
+    jp_camc = Column(Float)
+    jp_lcom = Column(Float)
+    jp_mmac = Column(Float)
+    jp_nhd = Column(Float)
+    jp_scom = Column(Float)
+
+
